Move Mailer class under top level libs/mail.py
